Remove slice pointers from assim() and aggr()
The original slice parameters are still modified in place but now
the new size is returned and it's up to the caller to delete the
extra elements.

This still isn't idiomatic Go but is the most efficient.  Using
a return slice is significantly slower.  A pointer receiver for
a method may be the best compromise.